fix duplicate key; upgrade to geo-explorer beta 11; separate out inline component
