+++ package.json
@@ -16,7 +16,7 @@
         "@mui/joy": "^5.0.0-beta.49",
         "@mui/material": "6.1.8",
         "@mui/x-data-grid": "^7.28.3",
-        "@ncsa/geo-explorer": "1.0.0-beta.9",
+        "@ncsa/geo-explorer": "1.0.0-beta.11",
         "@reduxjs/toolkit": "^2.2.7",
         "@rjsf/mui": "^5.24.3",
         "@rjsf/utils": "^5.24.3",
